Isolate ownership flow compiler fixtures
Keep intentional unresolved and indirect ownership calls out of the complete cross-assembly call graph fixture catalog.
